+++ perl/FSM/Composition/Plan.pm
@@ -12,6 +12,7 @@ sub new ($class, %args) {
         top_name => $args{top_name},
         ports => $args{ports} || [],
         links => $args{links} || [],
+        resolved_links => $args{resolved_links} || [],
         nets => $args{nets} || [],
         instances => $args{instances} || [],
         raw_spec => $args{raw_spec},
@@ -22,6 +23,7 @@ sub lane ($self) { return $self->{lane} }
 sub top_name ($self) { return $self->{top_name} }
 sub ports ($self) { return $self->{ports} }
 sub links ($self) { return $self->{links} }
+sub resolved_links ($self) { return $self->{resolved_links} }
 sub nets ($self) { return $self->{nets} }
 sub instances ($self) { return $self->{instances} }
 sub raw_spec ($self) { return $self->{raw_spec} }